requirements.txt
asgiref==3.6.0
Django==3.2.18
django-cors-headers
djangorestframework==3.11.0
django-environ==0.4.5
django-jsonfield==1.4.0
djangorestframework==3.11.0
pytz==2023.3
sqlparse==0.4.3
gunicorn
typing-extensions==4.5.0